move audio format checks to reusable audio check utils
