+++ osu.Game/Rulesets/Edit/Checks/CheckSongFormat.cs
@@ -2,12 +2,9 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using ManagedBass;
-using osu.Framework.Audio.Callbacks;
 using osu.Game.Beatmaps;
-using osu.Game.Extensions;
 using osu.Game.Rulesets.Edit.Checks.Components;
 
 namespace osu.Game.Rulesets.Edit.Checks
@@ -36,28 +33,17 @@ namespace osu.Game.Rulesets.Edit.Checks
             if (beatmapSet == null) yield break;
             if (audioFile == null) yield break;
 
-            using (Stream data = context.WorkingBeatmap.GetStream(audioFile.File.GetStoragePath()))
+            var audioFormat = AudioCheckUtils.GetAudioFormatFromFile(context, context.Beatmap.Metadata.AudioFile);
+
+            // If the format is not supported by BASS
+            if (audioFormat == 0)
             {
-                if (data == null || data.Length <= 0) yield break;
-
-                var fileCallbacks = new FileCallbacks(new DataStreamFileProcedures(data));
-                int decodeStream = Bass.CreateStream(StreamSystem.NoBuffer, BassFlags.Decode, fileCallbacks.Callbacks, fileCallbacks.Handle);
-
-                // If the format is not supported by BASS
-                if (decodeStream == 0)
-                {
-                    yield return new IssueTemplateFormatUnsupported(this).Create(audioFile.Filename);
-
-                    yield break;
-                }
-
-                var audioInfo = Bass.ChannelGetInfo(decodeStream);
-
-                if (!allowedFormats.Contains(audioInfo.ChannelType))
-                    yield return new IssueTemplateIncorrectFormat(this).Create(audioFile.Filename);
-
-                Bass.StreamFree(decodeStream);
+                yield return new IssueTemplateFormatUnsupported(this).Create(audioFile.Filename);
+                yield break;
             }
+
+            if (!allowedFormats.Contains(audioFormat))
+                yield return new IssueTemplateIncorrectFormat(this).Create(audioFile.Filename);
         }
 
         public class IssueTemplateFormatUnsupported : IssueTemplate
